Remove AnimatedSimulation
This patch folds the functionality from AnimatedSimulation into Timeline.
