+++ sky/packages/sky/lib/animation.dart
@@ -7,7 +7,6 @@
 /// This library depends only on core Dart libraries and the `newton` package.
 library animation;
 
-export 'src/animation/animated_simulation.dart';
 export 'src/animation/animated_value.dart';
 export 'src/animation/animation_performance.dart';
 export 'src/animation/clamped_simulation.dart';
@@ -15,4 +14,5 @@ export 'src/animation/curves.dart';
 export 'src/animation/forces.dart';
 export 'src/animation/scheduler.dart';
 export 'src/animation/scroll_behavior.dart';
-export 'src/animation/timeline.dart';
+export 'src/animation/simulation_stepper.dart';
+export 'src/animation/ticker.dart';
